Configurable hook secret and centralised package globals
Hook secret is now read from the configuration file.
Package globals and types that are used in multiple files have been
moved to pkg.go, which also contains the package docstring.
